requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via django
attrs==23.1.0
    # via
    #   jsonschema
    #   referencing
django==4.2.5
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   djangorestframework
    #   drf-spectacular
django-cors-headers==4.2.0
    # via -r requirements.in
django-environ==0.11.2
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
django-filter==23.3
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   drf-spectacular
drf-spectacular[sidedcar]==0.26.5
    # via -r requirements.in
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.19.1
    # via drf-spectacular
jsonschema-specifications==2023.7.1
    # via jsonschema
psycopg2==2.9.9
    # via -r requirements.in
pytz==2023.3.post1
    # via djangorestframework
pyyaml==6.0.1
    # via drf-spectacular
referencing==0.30.2
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.10.3
    # via
    #   jsonschema
    #   referencing
sqlparse==0.4.4
    # via django
uritemplate==4.1.1
    # via drf-spectacular
